Rename live module to old_live
